Audio basically working, without testing Trying to get file name and size from the net
