+++ app/src/main/java/com/inshodesign/bossrss/RSSItemsFragment.java
@@ -1,6 +1,7 @@
 
 package com.inshodesign.bossrss;
 
+        import android.content.Context;
         import android.media.AudioManager;
         import android.media.MediaPlayer;
         import android.os.Bundle;
@@ -43,7 +44,7 @@ package com.inshodesign.bossrss;
  * Created by JClassic on 3/5/2017.
  */
 
-public class RSSItemsFragment extends Fragment implements MediaPlayer.OnPreparedListener, MediaController.MediaPlayerControl, TouchEv{
+public class RSSItemsFragment extends Fragment {
 
     private RecyclerView mRecyclerView;
     RSSContentsAdapter mAdapter;
@@ -51,24 +52,21 @@ public class RSSItemsFragment extends Fragment implements MediaPlayer.OnPrepared
     private RxBus _rxBus = new RxBus();
     private long mLastClickTime = 0;
 
+    OnFragmentInteractionListener mCallback;
 //    private StreamAudioPlayer mStreamAudioPlayer;
 //    private byte[] mBuffer = new byte[1024];
 //    private RxMediaPlayer mediaPlayer;
 //    private MediaPlayer mp;
 
 
-    /****/
-    private LinearLayout anchorlayout;
-    private MediaPlayer mediaPlayer;
-    private MediaController mediaController;
-    private Handler handler = new Handler();
+
 
     @Override
     public View onCreateView(LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_list, container, false);
         mRecyclerView = (RecyclerView) view.findViewById(R.id.recycler);
-        anchorlayout = (LinearLayout) view.findViewById(R.id.anchorlayout);
+
         return view;
     }
 
@@ -127,8 +125,8 @@ public class RSSItemsFragment extends Fragment implements MediaPlayer.OnPrepared
 //                            }
 //
 //                            RxMediaPlayer.play(RxMediaPlayer.from(audioStream.getPath()));
-
-                            runMediaPlayer(audioStream);
+                            mCallback.playAudio(audioStream);
+//                            runMediaPlayer(audioStream);
 
                         }
                     }
@@ -160,106 +158,16 @@ public class RSSItemsFragment extends Fragment implements MediaPlayer.OnPrepared
         return rssList;
     }
 
-    private void runMediaPlayer(AudioStream audioStream) {
-
-
-        mediaPlayer = new MediaPlayer();
-        mediaPlayer.setOnPreparedListener(this);
-        mediaController = new MediaController(getActivity());
+    @Override
+    public void onAttach(Context context) {
+        super.onAttach(context);
         try {
-            mediaPlayer.setDataSource(audioStream.getPath());
-            mediaPlayer.prepare();
-            mediaPlayer.start();
-        } catch (IOException e) {
-            Log.e("TEST", "Could not open file " + audioStream.getPath() + " for playback.", e);
+            mCallback = (OnFragmentInteractionListener) context;
+        } catch (ClassCastException e) {
+            throw new ClassCastException(context.toString()
+                    + " must implement OnHeadlineSelectedListener");
         }
     }
 
 
-    @Override
-    public void onStop() {
-       super.onStop();
-        mediaController.hide();
-        mediaPlayer.stop();
-        mediaPlayer.release();
-    }
-
-    /** Media Player Stuff **/
-
-
-//
-//    @Override
-//    private void onStop() {
-//        super.onStop();
-//        mediaController.hide();
-//        mediaPlayer.stop();
-//        mediaPlayer.release();
-//    }
-
-//    @Override
-//    public boolean onTouchEvent(MotionEvent event) {
-//        //the MediaController will hide after 3 seconds - tap the screen to make it appear again
-//        mediaController.show();
-//        return false;
-//    }
-
-    //--MediaPlayerControl methods----------------------------------------------------
-    public void start() {
-        mediaPlayer.start();
-    }
-
-    public void pause() {
-        mediaPlayer.pause();
-    }
-
-    public int getDuration() {
-        return mediaPlayer.getDuration();
-    }
-
-    public int getCurrentPosition() {
-        return mediaPlayer.getCurrentPosition();
-    }
-
-    public void seekTo(int i) {
-        mediaPlayer.seekTo(i);
-    }
-
-    public boolean isPlaying() {
-        return mediaPlayer.isPlaying();
-    }
-
-    public int getBufferPercentage() {
-        return 0;
-    }
-
-    public boolean canPause() {
-        return true;
-    }
-
-    public boolean canSeekBackward() {
-        return true;
-    }
-
-    public boolean canSeekForward() {
-        return true;
-    }
-    //--------------------------------------------------------------------------------
-
-    public void onPrepared(MediaPlayer mediaPlayer) {
-        Log.d("Test", "onPrepared");
-        mediaController.setMediaPlayer(this);
-        mediaController.setAnchorView(anchorlayout);
-
-        handler.post(new Runnable() {
-            public void run() {
-                mediaController.setEnabled(true);
-                mediaController.show();
-            }
-        });
-    }
-
-    @Override
-    public int getAudioSessionId() {
-        return 0;
-    }
 }
