+++ app/src/main/res/values/strings.xml
@@ -15,5 +15,14 @@
     <string name="menubuttonadd">Add RSS</string>
     <string name="menubuttonfacebook">Share RSS</string>
 
+    <string-array name="testfeeds" >
+        <item>http://www.thestar.com/feeds.topstories.rss</item>
+        <item>https://www.espn.com/espnradio/feeds/rss/podcast.xml/_/id/2406595</item>
+        <item>http://pitchfork.com/rss/reviews/albums/</item>
+        <item>http://feeds.bbci.co.uk/news/video_and_audio/news_front_page/rss.xml?edition=uk</item>
+        <item>http://timesofindia.indiatimes.com/rssfeedstopstories.cms</item>
+        <item></item>
+
+    </string-array>
 
 </resources>
